pyproject.toml
[tool.poetry]
name = "w3cc-tradingbot"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
poetry = "^1.2.1"
pandas = "^1.5.0"
numpy = "^1.23.3"
ccxt = "^1.94.2"
asyncio = "^3.4.3"
python-dotenv = "^0.21.0"
simplejson = "^3.17.6"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
